fitwell_backend/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
